Locate local metadata outside of AnnotationHub

Locate local metadata outside of AnnotationHub

local_metadata(local_path = "/dcl01/lieber/ajaffe/lab/GenomicState/data-raw")

Arguments

local_path

A character(1) pointing to where the data is stored locally

Value

The AnnotationHub metadata data.frame() for the data in this package with RDataPath updated to point to the local_path. It includes an addtional column called loadCode which you can evaluate with eval(parse(text = entry)).

See also

Author

Leonardo Collado-Torres

Examples

## Get the local metadata
meta <- local_metadata()

## Subset to the data of interest, lets say hg19 TxDb for v31
interest <- subset(meta, RDataClass == "TxDb" & Tags == "Gencode:v31:hg19")

## Inspect the result
interest
#>                                       Title
#> 49 TxDb for Gencode v31 on hg19 coordinates
#>                                                                                                                                                            Description
#> 49 Gencode v31 TxDb object on hg19 coordinates. This is useful for packages such as derfinder and derfinderPlot. For more information, check the GenomicState package.
#>    BiocVersion Genome SourceType
#> 49         3.1 GRCh37        GTF
#>                                                                                                                SourceUrl
#> 49 ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_31/GRCh37_mapping/gencode.v31lift37.annotation.gtf.gz
#>    SourceVersion      Species TaxonomyId Coordinate_1_based DataProvider
#> 49   Oct 03 2019 Homo sapiens       9606               TRUE      GENCODE
#>                                         Maintainer RDataClass DispatchClass
#> 49 Leonardo Collado-Torres <lcolladotor@gmail.com>       TxDb    SQLiteFile
#>                                                                              RDataPath
#> 49 /dcl01/lieber/ajaffe/lab/GenomicState/data-raw/gencode/gencode_v31_hg19_txdb.sqlite
#>                Tags
#> 49 Gencode:v31:hg19
#>                                                                                                                                 loadCode
#> 49 gencode_v31_hg19_txdb <- AnnotationDbi::loadDb("/dcl01/lieber/ajaffe/lab/GenomicState/data-raw/gencode/gencode_v31_hg19_txdb.sqlite")

## Next you can load the data
if (file.exists(interest$RDataPath)) {
    ## This only works at JHPCE
    eval(parse(text = interest$loadCode))

    ## Explore the loaded object (would be gencode_v31_hg19_txdb in this case)
    gencode_v31_hg19_txdb
}